apps/api/pyproject.toml
[tool.poetry]
name = "orchestrai-api"
version = "0.1.0"
description = "FastAPI services for OrchestrAI"
authors = ["Mac Name <mac@macanderson.com>"]
readme = "README.md"
packages = [{ include = "api", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = { extras = ["standard"], version = "^0.34.2" }
fastapi-jwt-auth = "^0.5.0"
beautifulsoup4 = "^4.13.4"
openai = "^1.81.0"
langchain-community = "^0.3.24"
langchain-openai = "^0.3.18"
qdrant-client = "^1.14.2"
urllib3 = "^2.4.0"
sentence-transformers = "^4.1.0"
supabase = "^2.15.1"
python-jose = "^3.4.0"
bcrypt = "^4.3.0"
tomlkit = "^0.13.2"
pydantic = {extras = ["email"], version = "^2.11.5"}
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
start = "src.api.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"